+++ fpscompress-template-1.21.11/src/main/java/com/mukulramesh/fpscompress/gui/PreFabStatusMenu.java
@@ -96,9 +96,6 @@ public class PreFabStatusMenu extends AbstractContainerMenu {
                             prefabEntity.getSimulationRequiredTicks(),
                             prefabEntity.getCurrentDisplayMode(),
                             prefabEntity.getFocusedResourceId(),
-                            prefabEntity.getAutoNormalizedTicks(),
-                            prefabEntity.getUseAutoNormalize(),
-                            prefabEntity.getAutoNormalizedDisplayMode(),
                             prefabEntity.getPrefabName()
                         )
                     );